+++ src/targets/gpu/kernels/include/migraphx/kernels/basic_ops.hpp
@@ -34,7 +34,7 @@ struct id
 
 struct mean
 {
-    size_t item_num = 1;
+    int item_num = 1;
     template <class T>
     constexpr auto operator()(T x) const
     {
