+++ test/verify/test_batchnorm_3d_per_actv.cpp
@@ -10,11 +10,11 @@
 
 struct test_batchnorm_3d_per_actv : verify_program<test_batchnorm_3d_per_actv>
 {
-    const size_t d1       = 2;
-    const size_t d2       = 4;
-    const size_t d3       = 5;
-    const size_t channels = 2;
-    const size_t batches  = 3;
+    const int d1       = 2;
+    const int d2       = 4;
+    const int d3       = 5;
+    const int channels = 2;
+    const int batches  = 3;
 
     migraphx::program create_program() const
     {
